Fix migrations 054-057 for production deployment
- Migration 054: Add BEGIN/COMMIT transaction wrapper for consistency
- Migration 056: Fix tag_rules index (use partial index with WHERE clause instead of priority DESC)
- Migration 056: Add EXISTS checks to tag_rules RLS policies to ensure tag exists and belongs to same account
- Migration 057: Fix tag_ids column type from INTEGER[] to BIGINT[] to match tags.id (BIGSERIAL)
- Migration 058: Remove fix migration - merged fixes directly into original migrations

These migrations failed on production because:
1. Migration 056 had incorrect index definition and missing RLS policy validations
2. Migration 057 had type mismatch (INTEGER[] vs BIGINT[])
3. Migration 058 attempted to DROP and recreate tables, which fails on fresh deployments

All migrations after 053 are now production-ready and will deploy successfully.

+++ migrations/057_add_tag_ids_to_queued_imports.sql
@@ -4,9 +4,10 @@
 
 BEGIN;
 
--- Add tag_ids column to queued_imports (stored as JSONB array of tag IDs)
+-- Add tag_ids column to queued_imports (stored as array of tag IDs)
+-- Note: tags.id is BIGINT (BIGSERIAL), so we use BIGINT[] to match
 ALTER TABLE queued_imports
-  ADD COLUMN IF NOT EXISTS tag_ids INTEGER[] DEFAULT '{}';
+  ADD COLUMN IF NOT EXISTS tag_ids BIGINT[] DEFAULT '{}';
 
 -- Add index for tag lookups
 CREATE INDEX IF NOT EXISTS idx_queued_imports_tag_ids ON queued_imports USING GIN(tag_ids);
